requirements.txt
Flask==1.0.2
pymongo==3.6.1
numpy==1.18.0
Pillow
tensorflow
matplotlib
glob2